@@ -1,5 +1,13 @@
+<script lang="ts">
+	interface Props {
+		children?: import('svelte').Snippet;
+	}
+
+	let { children }: Props = $props();
+</script>
+
 <main class="container mx-auto px-4 py-8 md:px-8">
 	<div class="grid grid-cols-1 gap-8 lg:grid-cols-3">
-		<slot></slot>
+		{@render children?.()}
 	</div>
 </main>
